+++ src/components/SignIn/SignIn.styles.scss
@@ -5,7 +5,7 @@
   left: 28%;
   transform: translateY(-50%);
   z-index: 0;
-  height: 65%;
+  height: 60%;
   width: 30%;
   background-color: rgba(255, 255, 255, 0.918);
   border-radius: 10px;
@@ -14,17 +14,15 @@
   margin: auto;
   text-align: center;
   
-
   .login-wrapper {
     position: relative;
-    top: 50%;
-    transform: translateY(-50%);
+    top: 5%;
+    // transform: translateY(-50%);
   }
-
 };
 
 .user-inputs-wrappers {
-  margin-top: 11px;
+  margin-top: 10px;
 
   .user-inputs {
     width: 60%;
@@ -32,18 +30,18 @@
 };
 
 .login-button {
-  margin-top: 28px;
-  margin-bottom: 28px;
+  margin-top: 22px;
+  margin-bottom: 10px;
+  .login-text {
+    color: white;
+  }
 };
-
-.login-text {
-  color: white;
-}
 
 .signup-forgot-text{
   color: blue;
   font-size: 12px;
-  margin-bottom: 8px;
+  margin-bottom: 14px;
+  // display: block;
 
   &:hover{
     // cursor: pointer;
@@ -54,6 +52,6 @@
 @media only screen and (max-width: 980px) {
   .SignIn {
   left: 0;
-  width: 60%;
+  width: 80%;
   }
 };
