--- src/components/SignUp/SignUp.component.jsx
@@ -1,47 +1,0 @@
-import React from 'react';
-import "./SignUp.styles.scss";
-
-import { Link } from "react-router-dom";
-
-import TextField from '@material-ui/core/TextField';
-import Button from '@material-ui/core/Button';
-
-export const SignUp = () => {
-  return (
-    <div>
-      ewf
-
-      {/* <div className="login-wrapper">
-        <h1>Log In</h1>
-        <form noValidate autoComplete="off">
-          <div className="user-inputs-wrappers">
-            <TextField label="Email" className="user-inputs" type="email" />
-          </div>
-
-          <div className="user-inputs-wrappers">
-            <TextField label="Password" className="user-inputs" type="password" />
-          </div>
-
-          <div className="login-button">
-            <Button variant="contained" color="primary" >
-              <span className="login-text">Log In</span>
-            </Button>
-          </div>
-
-          <Link to="/signup">
-            <p className="signup-forgot-text">Don't have an account? Sign up here!</p>
-          </Link>
-
-          <Link to="/">
-            <p className="signup-forgot-text">I forgot my password</p>
-          </Link>
-        </form>
-      </div> */}
-
-
-    </div>
-  )
-
-
-
-}
