+++ src/pages/Landing/Landing.styles.scss
@@ -1,0 +1,53 @@
+.Landing {
+  height: 100vh;
+  position: relative;
+  font-family: Gill Sans, sans-serif;
+
+  .navbar {
+    position: relative;
+    background-color: white;
+    z-index: 1;
+    height: 60px;
+    box-shadow: 0 0 9px rgba(0, 0, 0, 0.157);
+    position: sticky;
+    top: 0;
+    padding-left: 80px;
+
+    // .navbar-logo {
+    //   // display: inline-block;
+    //   // border: 1px black solid;
+    //   height: 80%;
+    //   width: 100px;
+    //   background-image: url("../../assets/img/ichatLogo.png");
+    //   background-size: cover;
+    // }
+
+    .navbar-items {
+      vertical-align: top;
+      margin-right: 30px;
+      display: inline-block;
+      position: relative;
+      cursor: pointer;
+      line-height: 1;
+      top: 50%;
+      transform: translateY(-50%);
+
+      &:hover {
+        text-decoration: underline;
+      }
+
+      p:hover {
+        transform: scale(1.08);
+      }
+    }
+  }
+
+  .Landing-background-pic {
+    background-image: url("../../assets/img/LandingPageBackground.jpg");
+    background-size: cover;
+    background-position: bottom;
+    // height: 90vh;
+    height: 91.8vh;
+    min-height: 500px;
+  };
+};
