--- src/pages/Landing/SignInLanding.styles.scss
@@ -1,104 +1,0 @@
-.Landing {
-  height: 2000px;
-  position: relative;
-  font-family: Gill Sans, sans-serif;
-
-  .navbar {
-    position: relative;
-    background-color: white;
-    z-index: 1;
-    height: 60px;
-    box-shadow: 0 0 9px rgba(0, 0, 0, 0.157);
-    position: sticky;
-    top: 0;
-    padding-left: 80px;
-
-    .navbar-items {
-      margin-right: 30px;
-      display: inline-block;
-      position: relative;
-      cursor: pointer;
-      line-height: 1;
-      top: 50%;
-      transform: translateY(-50%);
-
-      &:hover {
-        text-decoration: underline;
-      }
-
-      p:hover {
-        transform: scale(1.08);
-      }
-    }
-  }
-
-  .Landing-background-pic {
-    background-image: url("../../assets/img/LandingPageBackground.jpg");
-    background-size: cover;
-    background-position: bottom;
-    height: 90vh;
-    min-height: 500px;
-  };
-
-  .SignIn {
-    font-family: "Comic Sans MS", cursive, sans-serif;
-    position: relative;
-    top: 50%;
-    left: 28%;
-    transform: translateY(-50%);
-    z-index: 0;
-    height: 65%;
-    width: 30%;
-    background-color: rgba(255, 255, 255, 0.918);
-    border-radius: 10px;
-    max-width: 400px;
-    min-width: 280px;
-    margin: auto;
-    text-align: center;
-    
-  
-    .login-wrapper {
-      position: relative;
-      top: 50%;
-      transform: translateY(-50%);
-    }
-  
-  };
-  
-  .user-inputs-wrappers {
-    margin-top: 11px;
-  
-    .user-inputs {
-      width: 60%;
-    }
-  };
-  
-  .login-button {
-    margin-top: 28px;
-    margin-bottom: 28px;
-  };
-  
-  .login-text {
-    color: white;
-  }
-  
-  .signup-forgot-text{
-    color: blue;
-    font-size: 12px;
-    margin-bottom: 8px;
-  
-    &:hover{
-      // cursor: pointer;
-      text-decoration: underline;
-    }
-  }
-  
-  @media only screen and (max-width: 980px) {
-    .SignIn {
-    left: 0;
-    width: 60%;
-    }
-  };
-  
-  
-}
